# bpo-37723: fix performance regression on regular expression parsing