--- a/Misc/ACKS
+++ b/Misc/ACKS
@@ -1707,6 +1707,7 @@ Michael Urman
 Hector Urtubia
 Lukas Vacek
 Ville Vainio
+Yann Vaginay
 Andi Vajda
 Case Van Horsen
 John Mark Vandenberg