--- a/Misc/NEWS.d/next/Library/2019-07-31-16-49-01.bpo-37723.zq6tw8.rst
+++ b/Misc/NEWS.d/next/Library/2019-07-31-16-49-01.bpo-37723.zq6tw8.rst
@@ -1,1 +1,2 @@
-Fix performance regression on regular expression parsing
+Fix performance regression on regular expression parsing with huge
+character sets. Patch by Yann Vaginay.