@yannvgn 2019-07-31
I have made the requested changes; please review again